+++ previous-versions/banking/src/sdk/sdk.ts
@@ -56,9 +56,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.32.0";
-    genVersion = "2.142.2";
-    userAgent = "speakeasy-sdk/typescript 0.32.0 2.142.2 3.0.0 @codat/banking";
+    sdkVersion = "0.32.1";
+    genVersion = "2.143.2";
+    userAgent = "speakeasy-sdk/typescript 0.32.1 2.143.2 3.0.0 @codat/banking";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);